# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API backend de la plateforme RAYCARD (wallet Mobile Money + cartes Visa virtuelles).",
        "title": "RAYCARD API",
        "contact": {
            "name": "RAYCARD — HR-Skills SARL"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/connexion": {
            "post": {
                "description": "Vérifie email + mot de passe et déclenche la 2FA obligatoire : envoie un code par email et renvoie un ticket à présenter avec ce code sur /auth/connexion/verifier-code. Aucun token de session n'est émis ici.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Connexion (étape 1/2)",
                "parameters": [
                    {
                        "description": "Identifiants",
                        "name": "connexion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ConnexionRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.ConnexionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "identifiants invalides",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/connexion-google": {
            "post": {
                "description": "Vérifie l'ID token Google (crée le compte automatiquement s'il n'existe pas encore, ou lie ce compte Google à un compte existant de même email si Google confirme l'email vérifié), puis déclenche la 2FA obligatoire comme pour /auth/connexion : envoie un code par email et renvoie un ticket à présenter sur /auth/connexion/verifier-code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Connexion via Google (étape 1/2)",
                "parameters": [
                    {
                        "description": "ID token Google, téléphone et pays (utilisés seulement à la création)",
                        "name": "connexion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ConnexionGoogleRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.ConnexionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide, ou téléphone/pays manquants pour une première connexion",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "id token invalide, ou email non vérifié pour lier un compte existant",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/connexion/verifier-code": {
            "post": {
                "description": "Échange le ticket obtenu à l'étape 1 et le code reçu par email contre une session complète (access + refresh token). 5 tentatives maximum ; au-delà, le ticket est définitivement invalidé et une alerte de sécurité est envoyée. Une connexion réussie déclenche aussi une notification par email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Connexion (étape 2/2) — vérification du code",
                "parameters": [
                    {
                        "description": "Ticket et code reçu par email",
                        "name": "verification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.VerifierCode2FARequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.SessionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "ticket ou code invalide, expiré, ou tentatives épuisées",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/deconnexion": {
            "post": {
                "description": "Révoque le refresh token fourni (déconnexion idempotente : ne renvoie pas d'erreur si déjà invalide)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Déconnexion",
                "parameters": [
                    {
                        "description": "Refresh token à révoquer",
                        "name": "deconnexion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RafraichirRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/empreinte/appareils": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Associe la clé publique générée par l'appareil (après déverrouillage biométrique) à l'utilisateur authentifié. L'empreinte elle-même et la clé privée ne quittent jamais l'appareil.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Enregistrement d'un appareil pour la connexion par empreinte",
                "parameters": [
                    {
                        "description": "Clé publique de l'appareil",
                        "name": "appareil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.EnregistrerAppareilRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.AppareilResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide ou clé publique mal formée",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/empreinte/appareils/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalide définitivement un appareil (perte, vol, ou l'utilisateur ne veut plus l'utiliser).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Révocation d'un appareil enregistré pour l'empreinte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de l'appareil",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "appareil introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/empreinte/challenge": {
            "post": {
                "description": "Émet un nonce à signer avec la clé privée de l'appareil désigné (déverrouillée par empreinte côté client).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Demande d'un challenge pour la connexion par empreinte",
                "parameters": [
                    {
                        "description": "Appareil concerné",
                        "name": "demande",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.DemanderChallengeEmpreinteRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.ChallengeEmpreinteResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "appareil inconnu ou révoqué",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/auth/empreinte/verifier": {
            "post": {
                "description": "Échange un challenge signé par la clé privée de l'appareil contre une session complète. Contrairement aux autres méthodes de connexion, aucun code par email n'est requis : l'appareil et l'empreinte qui l'a déverrouillé constituent déjà deux facteurs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Connexion par empreinte digitale",
                "parameters": [
                    {
                        "description": "Challenge signé",
                        "name": "verification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.VerifierEmpreinteRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.SessionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "401": {
                        "description": "challenge invalide, expiré, ou signature incorrecte",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
...
        "/auth/mot-de-passe-oublie": {
            "post": {
                "description": "Envoie un code par email si un compte existe pour cet email. Répond toujours le même succès générique, que le compte existe ou non (évite l'énumération de comptes).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"1. Client - Auth\""
                ],
                "summary": "Demande de réinitialisation de mot de passe",
                "parameters": [
                    {
                        "description": "Email du compte",
                        "name": "demande",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.DemanderReinitialisationRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "corps de requête invalide",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.ConnexionGoogleRequestDTO": {
            "type": "object",
            "required": [
                "id_token",
                "pays_code",
                "telephone"
            ],
            "properties": {
                "id_token": {
                    "type": "string"
                },
                "pays_code": {
                    "type": "string",
                    "example": "CI"
                },
                "telephone": {
                    "type": "string",
                    "example": "+2250700000000"
                }
            }
        },
        "auth.ConnexionRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "mot_de_passe"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "awa.kone@example.com"
                },
                "mot_de_passe": {
                    "type": "string",
                    "example": "motdepasse123"
                }
            }
        },
        "auth.ConnexionResponseDTO": {
            "type": "object",
            "properties": {
                "expire_dans_sec": {
                    "type": "integer"
                },
                "ticket": {
                    "type": "string"
                }
            }
        },
...
        "commun.ErreurDTO": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string",
                    "example": "un compte existe déjà avec cet email"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}